requirements.txt
absl-py==2.0.0
aniso8601==9.0.1
asgiref==3.4.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
blinker==1.7.0
bson==0.5.10
cachetools==5.3.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.1.7
contourpy==1.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
cycler==0.12.1
dataclasses==0.6
DateTime==5.2
declxml==1.1.3
decorator==5.1.1
Deprecated==1.2.14
Django==3.1.5
django-cors-headers==3.4.0
django-currentuser==0.6.1
djangorestframework==3.11.0
dnspython==2.4.2
dpcpp-cpp-rt==2023.2.0
drf-yasg==1.17.1
executing==2.0.1
filelock==3.13.1
Flask==3.0.0
flask-restplus==0.13.0
fonttools==4.43.1
fsspec==2023.10.0
google-auth==2.23.3
google-auth-oauthlib==1.0.0
grpcio==1.59.2
gunicorn==20.1.0
huggingface-hub==0.17.3
idna==3.3
imageio==2.31.6
imgaug==0.4.0
importlib-metadata==6.8.0
importlib-resources==6.1.0
inflection==0.5.1
intel-cmplr-lib-rt==2023.2.0
intel-cmplr-lic-rt==2023.2.0
intel-opencl-rt==2023.2.0
intel-openmp==2023.2.0
interchange==2021.0.4
ipython==8.12.3
iso8601==2.1.0
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.2.0
jsons==1.6.3
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
labelImg==1.8.6
lazy_loader==0.3
lxml==4.9.3
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mkl==2023.2.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.3.0
neo4j==5.12.0
neoapi==3.6.4
neobolt==1.7.17
neomodel==5.2.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opencv-python==4.8.1.78
packaging==21.0
pandas==1.3.4
pansi==2020.7.3
parso==0.8.3
passlib==1.7.4
pickle5==0.0.11
pickleshare==0.7.5
Pillow==8.4.0
pkgutil_resolve_name==1.3.10
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.6
pure-eval==0.2.2
py-cpuinfo==9.0.0
py2neo==2021.2.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.20
Pygments==2.16.1
pymongo==4.0.1
pyparsing==3.0.1
pyserial==3.5
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.4.1
redis==4.3.4
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.6
rsa==4.9
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
sacremoses==0.1.1
safetensors==0.4.0
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.9.2
seaborn==0.13.0
serial==0.0.97
serial-tool==0.0.1
shapely==2.0.2
six==1.16.0
sqlparse==0.4.2
stack-data==0.6.3
sympy==1.12
tbb==2021.10.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.7.10
tkintertable==1.3.3
tokenizers==0.14.1
torch==1.7.0
torchvision==0.8.1
tqdm==4.66.1
traitlets==5.13.0
transformers==4.34.1
triton==2.1.0
typing==3.7.4.3
typing_extensions==4.8.0
typish==1.9.3
ultralytics==8.0.215
uritemplate==4.1.1
urllib3==1.26.7
wcwidth==0.2.9
Werkzeug==3.0.1
wrapt==1.15.0
XlsxWriter==3.0.3
xmltodict==0.13.0
zipp==3.17.0
zope.interface==6.1
